updated scripts, data and figure

+++ analysis/eval_time/ground_state/get_times_NGPU.py
@@ -1,43 +1,49 @@
 import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
+import os
 
-datapath = ""
-name_times = ['Building', 'Simulation', 'neuron_Update_time',
-'poisson_generator_time', 'GetSpike_time', 'NestedLoop_time', 'SpikeBufferUpdate_time',
-'SpikeReset_time', 'copy_ext_spike_time', 'SendExternalSpike_time', 'SendSpikeToRemote_time',
-'RecvSpikeFromRemote_time', 'ExternalSpikeReset_time']
+if os.path.isfile("./raw_times.csv"):
+    df = pd.read_csv("raw_times.csv", index_col=0)
+else:
+    datapath = "raw_times"
+    name_times = ['Building', 'Simulation', 'neuron_Update_time',
+    'poisson_generator_time', 'GetSpike_time', 'NestedLoop_time', 'SpikeBufferUpdate_time',
+    'SpikeReset_time', 'copy_ext_spike_time', 'SendExternalSpike_time', 'SendSpikeToRemote_time',
+    'RecvSpikeFromRemote_time', 'ExternalSpikeReset_time']
 
-label = []
-times = []
+    label = []
+    times = []
 
-for filename in name_times:
-    fn = datapath+"/all_"+filename+".dat"
-    file = np.loadtxt(fn)
-    #print("File "+filename+": ", file[1::2])
-    times.append(file[1::2])
-    #compute names
-    if (filename == "Building"):
-        label.append("Building time")
-    elif (filename == "Simulation"):
-        label.append("Simulation time")
-    else:
-        label.append(filename)
+    for filename in name_times:
+        fn = datapath+"/all_"+filename+".dat"
+        file = np.loadtxt(fn)
+        #print("File "+filename+": ", file[1::2])
+        if filename=="Building":
+            times.append(file[1::2])
+        else:
+            times.append(file[1::2] - file[::2])
+        #compute names
+        if (filename == "Building"):
+            label.append("Building time")
+        elif (filename == "Simulation"):
+            label.append("Simulation time")
+        else:
+            label.append(filename)
 
-data = {}
-for i in range(len(label)):
-    data[label[i]] = times[i]
+    data = {}
+    for i in range(len(label)):
+        data[label[i]] = times[i]
 
-df = pd.DataFrame(data)
-df.to_csv("raw_times.csv")
+    df = pd.DataFrame(data)
+    df.to_csv("raw_times.csv")
 
 bt = []
 st = []
 neuron = []
 poisson = []
-delivery = []
-communication = []
-collocation = []
+remote_spike = []
+local_spike = []
 other = []
 
 for i in range(10):
@@ -49,70 +55,57 @@ for i in range(10):
     max_st = max(run_i["Simulation time"])
     st.append(max_st)
     #extract the indexed that corresponds to the highest simulation times
-    max_st_idx = run_i.index[run_i["Simulation time"] == max_st].tolist()
+    #max_st_idx = run_i.index[run_i["Simulation time"] == max_st].tolist()
+    #extract all the indexed to average data from every compute node
+    max_st_idx = np.arange(i*32, (i+1)*32)
+    #print("Indici MAX:", max_st_idx)
     #sometimes there is more than one node that exibits the maximum value
     #in this case we average the results in order to show a single set of values per each node
-    if(len(max_st_idx) == 1):
-        coll = (run_i["RecvSpikeFromRemote_time"].loc[max_st_idx[0]] -
-                run_i["RecvSpikeFromRemote_MPI_time"].loc[max_st_idx[0]] +
-                run_i["SendSpikeToRemote_time"].loc[max_st_idx[0]] -
-                run_i["SendSpikeToRemote_MPI_time"].loc[max_st_idx[0]])
-        comm = (run_i["copy_ext_spike_time"].loc[max_st_idx[0]] +
-            run_i["SendExternalSpike_time"].loc[max_st_idx[0]] +
-            run_i["SendSpikeToRemote_time"].loc[max_st_idx[0]] +
-            run_i["RecvSpikeFromRemote_time"].loc[max_st_idx[0]] +
-            run_i["ExternalSpikeReset_time"].loc[max_st_idx[0]]) - coll
-        deliv = (run_i["GetSpike_time"].loc[max_st_idx[0]] +
-            run_i["NestedLoop_time"].loc[max_st_idx[0]] +
-            run_i["SpikeBufferUpdate_time"].loc[max_st_idx[0]] +
-            run_i["SpikeReset_time"].loc[max_st_idx[0]])
-        neuron.append(run_i["neuron_Update_time"].loc[max_st_idx[0]])
-        poisson.append(run_i["poisson_generator_time"].loc[max_st_idx[0]])
-        oth = (run_i["Simulation time"].loc[max_st_idx[0]] - coll - comm - deliv - neuron[-1] - poisson [-1])
-        collocation.append(coll)
-        communication.append(comm)
-        delivery.append(deliv)
-        other.append(oth)
-    else:
-        coll = (np.mean(run_i["RecvSpikeFromRemote_time"].loc[max_st_idx]) -
-                np.mean(run_i["RecvSpikeFromRemote_MPI_time"].loc[max_st_idx]) +
-                np.mean(run_i["SendSpikeToRemote_time"].loc[max_st_idx]) -
-                np.mean(run_i["SendSpikeToRemote_MPI_time"].loc[max_st_idx]))
-        comm = (np.mean(run_i["copy_ext_spike_time"].loc[max_st_idx]) +
-            np.mean(run_i["SendExternalSpike_time"].loc[max_st_idx]) +
-            np.mean(run_i["SendSpikeToRemote_time"].loc[max_st_idx]) +
-            np.mean(run_i["RecvSpikeFromRemote_time"].loc[max_st_idx]) +
-            np.mean(run_i["ExternalSpikeReset_time"].loc[max_st_idx])) - coll
-        deliv = (np.mean(run_i["GetSpike_time"].loc[max_st_idx]) +
-            np.mean(run_i["NestedLoop_time"].loc[max_st_idx]) +
-            np.mean(run_i["SpikeBufferUpdate_time"].loc[max_st_idx]) +
-            np.mean(run_i["SpikeReset_time"].loc[max_st_idx]))
-        neuron.append(np.mean(run_i["neuron_Update_time"].loc[max_st_idx]))
-        poisson.append(np.mean(run_i["poisson_generator_time"].loc[max_st_idx]))
-        oth = (np.mean(run_i["Simulation time"].loc[max_st_idx]) - coll - comm - deliv - neuron[-1] - poisson [-1])
-        communication.append(comm)
-        collocation.append(coll)
-        delivery.append(deliv)
-        other.append(oth)
+    
+    remote_spike_handling_delivery = (np.mean(run_i["copy_ext_spike_time"].loc[max_st_idx]) +
+        np.mean(run_i["SendExternalSpike_time"].loc[max_st_idx]) +
+        np.mean(run_i["SendSpikeToRemote_time"].loc[max_st_idx]) +
+        np.mean(run_i["RecvSpikeFromRemote_time"].loc[max_st_idx]) +
+        np.mean(run_i["ExternalSpikeReset_time"].loc[max_st_idx]))
+    local_spike_handling_delivery = (np.mean(run_i["GetSpike_time"].loc[max_st_idx]) +
+        np.mean(run_i["NestedLoop_time"].loc[max_st_idx]) +
+        np.mean(run_i["SpikeBufferUpdate_time"].loc[max_st_idx]) +
+        np.mean(run_i["SpikeReset_time"].loc[max_st_idx]))
+    neuron.append(np.mean(run_i["neuron_Update_time"].loc[max_st_idx]))
+    poisson.append(np.mean(run_i["poisson_generator_time"].loc[max_st_idx]))
+    remote_spike.append(remote_spike_handling_delivery)
+    local_spike.append(local_spike_handling_delivery)
+    oth = (np.mean(run_i["Simulation time"].loc[max_st_idx]) - 
+        remote_spike_handling_delivery - local_spike_handling_delivery - 
+        np.mean(run_i["neuron_Update_time"].loc[max_st_idx]) - 
+        np.mean(run_i["poisson_generator_time"].loc[max_st_idx]))
+    other.append(oth)
 
 print("\n\nAveraged results")
-print("Building time [s]    :", np.mean(bt), "+/-", np.std(bt))
-print("Simulation time [s]  :", np.mean(st), "+/-", np.std(st))
-print("Neuron dynamics [s]  :", np.mean(neuron), "+/-", np.std(neuron))
-print("Poisson generator [s]:", np.mean(poisson), "+/-", np.std(poisson))
-print("Communication [s]    :", np.mean(communication), "+/-", np.std(communication))
-print("Collocation [s]      :", np.mean(collocation), "+/-", np.std(collocation))
-print("Delivery [s]         :", np.mean(delivery), "+/-", np.std(delivery))
-print("Other [s]            :", np.mean(other), "+/-", np.std(other))
+print("Building time [s]     :", np.mean(bt), "+/-", np.std(bt))
+print("Simulation time [s]   :", np.mean(st), "+/-", np.std(st))
+print("Neuron dynamics [s]   :", np.mean(neuron), "+/-", np.std(neuron))
+print("Poisson generator [s] :", np.mean(poisson), "+/-", np.std(poisson))
+print("Remote spike (MPI) [s]:", np.mean(remote_spike), "+/-", np.std(remote_spike))
+print("Local spike [s]       :", np.mean(local_spike), "+/-", np.std(local_spike))
+print("Other [s]             :", np.mean(other), "+/-", np.std(other))
 
-mean = [np.mean(bt), np.mean(st), np.mean(neuron), np.mean(poisson), np.mean(communication), np.mean(collocation), np.mean(delivery), np.mean(other)]
-std = [np.std(bt), np.std(st), np.std(neuron), np.std(poisson), np.std(communication), np.std(collocation), np.std(delivery), np.std(other)]
+mean = [np.mean(bt), np.mean(st), np.mean(neuron), np.mean(poisson), np.mean(remote_spike), np.mean(local_spike), np.mean(other)]
+std = [np.std(bt), np.std(st), np.std(neuron), np.std(poisson), np.std(remote_spike), np.std(local_spike), np.std(other)]
 
 stat ={"mean": mean,"std": std}
 df_stat = pd.DataFrame(stat, index = ["Building time ", "Simulation time", "Neuron dynamics", "Poisson generator",
-        "Communication", "Collocation", "Delivery", "Other"])
+        "Remote spike", "Local spike", "Other"])
 df_stat.to_csv("ngpu_times_stat.csv")
 
 
-
-
+'''
+Averaged results
+Building time [s]     : 957.0352999999999 +/- 40.84523123217688
+Simulation time [s]   : 153.340226 +/- 8.626929746483624
+Neuron dynamics [s]   : 6.51167703125 +/- 0.02243970911894302
+Poisson generator [s] : 13.307004765625 +/- 0.05202330120969217
+Remote spike (MPI) [s]: 80.0172076053125 +/- 6.553334008125983
+Local spike [s]       : 38.6342937771875 +/- 2.9856293952018156
+Other [s]             : 14.868997383124995 +/- 0.42149699701596555
+'''
